[feature] extract ResumeService interface
